% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnat_api.R
\name{download_xnat_file}
\alias{download_xnat_file}
\title{Download XNAT file}
\usage{
download_xnat_file(conn, ...)
}
\arguments{
\item{conn}{The XNAT connection returned by a \code{xnat_connect} call}

\item{file_path}{Path to the file to be dowloaded}

\item{destfile}{Destination filename}

\item{prefix}{Prefix the file name with this (prevents
overwritting of same name files in case function is
used to download multiple scan types at once)}

\item{verbose}{Should progress be added to download?}

\item{error}{Should function error if download failed?}
}
\value{
Display path to the downloaded file
}
\description{
Download a single file from XNAT
}
\examples{
## file_path is retrieved using the get_scan_resources() function
xnat_central_conn <- xnat_connect('https://central.xnat.org', xnat_name="CENTRAL")
r <- get_scan_resources(xnat_central_conn,'CENTRAL_E00760')
\dontrun{download_nitrc_file(xnat_connect,r$URI[1])}
}
