% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnat_api.R
\name{get_scan_resources}
\alias{get_scan_resources}
\title{Get scan resources for a specific experiment ID}
\usage{
get_scan_resources(conn, ...)
}
\arguments{
\item{conn}{The XNAT connection returned by a \code{xnat_connect} call}

\item{...}{experiment_ID the experiment ID identifier, unique for each individual subject}
}
\description{
Get a full list of available resources for a specific experiment ID
}
\examples{
## Connect to XNAT CENTRAL
xnat_central_conn <- xnat_connect('https://central.xnat.org', xnat_name="CENTRAL")
get_scan_resources(xnat_central_conn,'CENTRAL_E00760')
}
