% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnat_api.R
\name{string2csv}
\alias{string2csv}
\title{Convert string to csv}
\usage{
string2csv(string)
}
\arguments{
\item{string}{input string}
}
\description{
Convert a string to csv format
}
\examples{
{
\dontrun{string2csv("col1,col2,col3\\ncell_1_1,cell_1_2,cell_1_3\\ncell_2_1,cell_2_2,cell_2_3")}
\dontrun{string2csv("c1,c2,c3\\n1,2,3")}
} 
}
